+++ NEWS.md
@@ -1,5 +1,13 @@
+# dv.listings 4.3.4-9009
+- Exported single listings include review information, if available
+
+# dv.listings 4.3.4-9008
+- Review functionality:
+  - Remove column change highlighting from rows that have been modified but restored afterwards
+
 # dv.listings 4.3.4-9007
-- Exported listings include review information, if available
+- Review functionality:
+  - Fix bug in column change highlight when known rows are removed
 
 # dv.listings 4.3.4-9006
 - Prevent header and body offsetting when dv.manager sidebar collapses/expands